Resources have unique names to avoid DCD warnings (#69)
At this moment, our created machines have all their resources with the
same names. This is normally not a problem as the resources are of
different types (server, volume, nic, ipblock, lan), but right now DCD
emits a warning complaining that they all have the same name. This is
expected from their side, so until something changes, we should append
the resource type to the resource name so we're consistent with DCD.


![image](https://github.com/ionos-cloud/cluster-api-provider-ionoscloud/assets/112630064/0f950cf9-ee91-466f-86c3-e415553886f7)