Support a custom CA bundle for Cloud API (#99)
When talking to a Cloud API instance using a certfificate signed by a CA
that is not trustworthy according to the system's root CA set, the
connection will be aborted by the client. This feature offers passing
the used CA as a trusted one.

This feature can also be useful for companies intercepting HTTPS
traffic. That includes a custom CA that needs to be trusted by all
devices whose traffic gets intercepted. This change allows setting such a
CA.

Not only a single CA can be given, but a whole bundle of them. The given
bundle is not used in addition to the system's root CA set, but instead
of it.

Similar features are present in many other projects. Some expect the CA
bundle to passed as base64-encoded input, others expect the plain PEM
without additional encoding on top. This change uses the latter approach.

This feature is unlikely to be used widely. As such the corresponding
key in the secret containing the Cloud API credentials is not rendered
in our provided template.