Allow FQDN as the control plane endpoint host
Until now, we assumed the control plane endpoint host to be an IP, but
that's an unnecessary restriction.

The template now consumes a new environment variable:
`CONTROL_PLANE_ENDPOINT_HOST`
It's optional and defaults to the value of the existing
`CONTROL_PLANE_ENDPOINT_IP` environment variable.

As the host is required to resolve to that IP, we'd actually not need it
anymore to be given explicitly. However, as we still render the kube-vip
manifest in the template and don't have anything nice there to do the
translation automatically, we keep the explicit IP for now.

The controller code is not aware of the `CONTROL_PLANE_ENDPOINT_IP`
value, though. It only sees the control plane endpoint host, which can
now be either an IP or an FQDN, so the controller must be able to
resolve the name to use the corresponding IP to find the matching
IP block and to find the correct IP failover group.

Unit tests for DNS resolving were not as convenient as I expected.
The stdlib tests don't provide any means to mock it, they even rely on
external network reachability to resolve some Google domain names.
They pointed to https://github.com/golang/go/issues/12503, though.
There's an interesting link to https://github.com/ncruces/go-dns,
showing how to inject custom behavior using a custom dialer, enabling
things such as caching. It's overkill for us, though, so our used
approach for unit testing is simpler.