:bug: Reduce failure risk (#107)
:warning: Breaking :warning: 

This PR introduces changes to resource names, which are created in IONOS
Cloud. This will lead to clusters being unable to be deleted. Therefore
clusters need to be deleted before using the new version.

**What is the purpose of this pull request/Why do we need it?**

Looks like api-server creates a label with a node name in its lease. A
label name has the limit of 63 characters. If the name for the node is
too long this fails with an error. As manifest names can be 253
characters long, we need to at least reduce the risk of failure

**Description of changes:**

* node names will now be the exact name, which is generated from the
machine deployment.
* Removed prefixes for resources to be consistent.


**Special notes for your reviewer:**

* We decided to remove the `k8s` prefix from resources.
* IONOS Cloud resources will still keep a prefix as the DCD will
complain if there are multiple resources with the same name

**Checklist:**
- [x] Unit Tests added
- [x] Includes
[emojis](https://github.com/kubernetes-sigs/kubebuilder-release-tools?tab=readme-ov-file#kubebuilder-project-versioning)
- [x] Local test successful