Allow FQDN as the control plane endpoint host (#108)
Until now, we assumed the control plane endpoint host to be an IP, but
that's an unnecessary restriction.

The template now consumes a new environment variable:
`CONTROL_PLANE_ENDPOINT_HOST`
It's optional and defaults to the value of the existing
`CONTROL_PLANE_ENDPOINT_IP` environment variable.

As the host is required to resolve to that IP, we'd actually not need it
anymore to be given explicitly. However, as we still render the kube-vip
manifest in the template and don't have anything nice there to do the
translation automatically, we keep the explicit IP for now.

The controller code is not aware of the `CONTROL_PLANE_ENDPOINT_IP`
value, though. It only sees the control plane endpoint host, which can
now be either an IP or an FQDN, so the controller must be able to
resolve the name to use the corresponding IP to find the matching IP
block and to find the correct IP failover group.