Lock `CurrentRequestByDatacenter` in cluster scope
Keeps the `api` package clean.

To scope the lock to a specific ICC, we make use of the keyed locks here
again, even though we don't need their ability to react on canceled
contexts. We therefore pass `context.Background` and ignore the error,
as none can occur.

In contrast to the previous approach, we lose the differentiation
between read-only and read-write locks, but even for writing the locks
are very short-lived, so it doesn't matter.