Add missing locker in cluster scope creations
(If we only had unit tests for the controllers...)

Locking `CurrentRequestByDatacenter` in the cluster scope is better than
doing in the machine scope. I had that first, but the code looked
strange and unrelated when suddenly patching the cluster after doing
things in the machine scope.

That's why the cluster scope contains a locker now. However, that
scope is used from the machine controller. It cannot (and should not)
access the cluster controller, so when building the cluster scope, the
machine controller needs to pass its own locker.
At the same time, the cluster controller also creates a cluster scope
and is now required to pass a locker, even though it doesn't actually
need one for its own work.

An alternative could be having a controller package-wide locker (global
variable)