:sparkles: Provide new controller for handling generic load balancer provisioning (#178)
**What is the purpose of this pull request/Why do we need it?**

This PR introduces a new controller to handle the setup of load
balancing for the control plane. This PR does not provide the actual
implementations as those will be done in separate PRs.

As this is intended to provide an interface to work on multiple
implementations at the same time, this branch will not be merged into
main to not have an intermediate state.

**Checklist:**
- [x] Unit Tests added
- [x] Includes
[emojis](https://github.com/kubernetes-sigs/kubebuilder-release-tools?tab=readme-ov-file#kubebuilder-project-versioning)